Support using a remote files structure
